Update libreg/src/rst.c to force double calculations of RST coefficients

+++ libreg/src/rst.c
@@ -174,7 +174,7 @@ static void regRstInitPII(struct reg_rst_pars  *pars,
 
     case 2:                                             // Algorithm 2 : Pure delay fraction 0.401 - 1.01 (not dead-beat)
 
-        pars->rst.r[0] = (3*a1 + c1 + d1 + 2*a1*c1 + 2*a1*d1 + a1*d2 - c1*d2 + a1*c1*d1 + 2)/(b0_b1*(a1 + 1)*(a1 + 1)) +
+        pars->rst.r[0] = (3*(double)a1 + c1 + d1 + 2*a1*c1 + 2*a1*d1 + a1*d2 - c1*d2 + a1*c1*d1 + 2)/(b0_b1*(a1 + 1)*(a1 + 1)) +
                          (b1*(c1 + 1)*(d1 + d2 + 1))/(b0_b1*b0_b1*(a1 + 1)) + (a1*(a1 - c1)*(a1*a1 - d1*a1 + d2))/((a1 + 1)*(a1 + 1)*(b1 - a1*b0));
 
         fprintf(stderr,"a1=%.8E\n",a1);
@@ -201,26 +201,26 @@ static void regRstInitPII(struct reg_rst_pars  *pars,
 
 
 
-        pars->rst.r[1] = (d2 + c1*d1 + 2*c1*d2 + 2*a1*a1*c1 + 2*a1*a1*d1 + a1*a1*d2 + 3*a1*a1 + a1*a1*c1*d1 - 1)/(b0_b1*(a1 + 1)*(a1 + 1)) -
+        pars->rst.r[1] = ((double)d2 + c1*d1 + 2*c1*d2 + 2*a1*a1*c1 + 2*a1*a1*d1 + a1*a1*d2 + 3*a1*a1 + a1*a1*c1*d1 - 1)/(b0_b1*(a1 + 1)*(a1 + 1)) -
                          (2*a1*(a1 - c1)*(a1*a1 - d1*a1 + d2))/((a1 + 1)*(a1 + 1)*(b1 - a1*b0)) + (b1*(a1 - 1)*(c1 + 1)*(d1 + d2 + 1))/(b0_b1*b0_b1*(a1 + 1));
 
-        pars->rst.r[2] = (a1*(a1 - c1*d2)*b0*b0 + a1*(2*a1 + d2 + c1*d1 - 1)*b0*b1 - a1*(c1 - a1 + d1 + 2)*b1*b1)/(b0_b1*b0_b1*(b1 - a1*b0));
+        pars->rst.r[2] = ((double)a1*(a1 - c1*d2)*b0*b0 + a1*(2*a1 + d2 + c1*d1 - 1)*b0*b1 - a1*(c1 - a1 + d1 + 2)*b1*b1)/(b0_b1*b0_b1*(b1 - a1*b0));
 
         pars->rst.s[0] = 1.0;
 
-        pars->rst.s[1] = (b0*b0*b1 + 2*b0*b1*b1 + b1*b1*b1)/(b0*b0*b0 + 2*b0*b0*b1 + b0*b1*b1) -
+        pars->rst.s[1] = ((double)b0*b0*b1 + 2*b0*b1*b1 + b1*b1*b1)/(b0*b0*b0 + 2*b0*b0*b1 + b0*b1*b1) -
                          (b1*(b1 - b0*c1)*(d2*b0*b0 - d1*b0*b1 + b1*b1))/(b0*b0_b1*b0_b1*(b1 - a1*b0)) - 2;
 
-        pars->rst.s[2] = (2*b1*(b1 - b0*c1)*(d2*b0*b0 - d1*b0*b1 + b1*b1))/(b0*b0_b1*b0_b1*(b1 - a1*b0)) -
+        pars->rst.s[2] = (2*(double)b1*(b1 - b0*c1)*(d2*b0*b0 - d1*b0*b1 + b1*b1))/(b0*b0_b1*b0_b1*(b1 - a1*b0)) -
                          (2*(b0*b0*b1 + 2*b0*b1*b1 + b1*b1*b1))/(b0*b0*b0 + 2*b0*b0*b1 + b0*b1*b1) + 1;
 
-        pars->rst.s[3] = (b0*b0*b1 + 2*b0*b1*b1 + b1*b1*b1)/(b0*b0*b0 + 2*b0*b0*b1 + b0*b1*b1) -
+        pars->rst.s[3] = ((double)b0*b0*b1 + 2*b0*b1*b1 + b1*b1*b1)/(b0*b0*b0 + 2*b0*b0*b1 + b0*b1*b1) -
                          (b1*(b1 - b0*c1)*(d2*b0*b0 - d1*b0*b1 + b1*b1))/(b0*b0_b1*b0_b1*(b1 - a1*b0));
 
         pars->rst.t[0] = 1.0 / b0_b1;
-        pars->rst.t[1] = (c1 + d1) / b0_b1;
-        pars->rst.t[2] = (c1*d1 + d2) / b0_b1;
-        pars->rst.t[3] = c1*d2 / b0_b1;
+        pars->rst.t[1] = ((double)c1 + d1) / b0_b1;
+        pars->rst.t[2] = ((double)c1*d1 + d2) / b0_b1;
+        pars->rst.t[3] = (double)c1*d2 / b0_b1;
 
         pars->rst.track_delay = 2.0 * pars->period;         // Track delay is about 2 periods (not dead-beat)
         break;
@@ -228,23 +228,23 @@ static void regRstInitPII(struct reg_rst_pars  *pars,
     case 3:                                            // Algorithm 3 : Pure delay fraction 1.01 - 1.401 (dead-beat)
 
         c2 = exp(-pars->period * TWO_PI * clbw3);
-        q1 = 2.0 - a1 + c1 + c2 + d1;
+        q1 = 2.0 - (double)a1 + c1 + c2 + d1;
 
-        pars->rst.r[0] = q1*(2.0 - a1) + d2 + c1*c2 + d1*(c1 + c2) + 2.0*a1 - 1.0;
-        pars->rst.r[1] = q1*(2.0*a1 - 1.0)  + c1*c2*d1 + d2*(c1 + c2) - a1;
-        pars->rst.r[2] = c1*c2*d2 - a1*q1;
+        pars->rst.r[0] = (double)q1*(2.0 - a1) + d2 + c1*c2 + d1*(c1 + c2) + 2.0*a1 - 1.0;
+        pars->rst.r[1] = (double)q1*(2.0*a1 - 1.0)  + c1*c2*d1 + d2*(c1 + c2) - a1;
+        pars->rst.r[2] = (double)c1*c2*d2 - a1*q1;
 
         pars->rst.s[0] = b0;
-        pars->rst.s[1] = b0*(q1 - 2.0) + b1;
-        pars->rst.s[2] = b1*(q1 - 2.0) - b0*(2.0*q1 - 1.0);
-        pars->rst.s[3] = b0*q1         - b1*(2.0*q1 - 1.0);
-        pars->rst.s[4] = b1*q1;
+        pars->rst.s[1] = (double)b0*(q1 - 2.0) + b1;
+        pars->rst.s[2] = (double)b1*(q1 - 2.0) - b0*(2.0*q1 - 1.0);
+        pars->rst.s[3] = (double)b0*q1         - b1*(2.0*q1 - 1.0);
+        pars->rst.s[4] = (double)b1*q1;
 
         pars->rst.t[0] = 1.0;
-        pars->rst.t[1] = c1 + c2 + d1;
-        pars->rst.t[2] = c1*c2 + d1*(c1 + c2) + d2;
-        pars->rst.t[3] = c1*c2*d1 + d2*(c1 + c2);
-        pars->rst.t[4] = c1*c2*d2;
+        pars->rst.t[1] = (double)c1 + c2 + d1;
+        pars->rst.t[2] = (double)c1*c2 + d1*(c1 + c2) + d2;
+        pars->rst.t[3] = (double)c1*c2*d1 + d2*(c1 + c2);
+        pars->rst.t[4] = (double)c1*c2*d2;
 
         pars->rst.track_delay = 2.0 * pars->period;         // Track delay is exactly 2 periods (dead-beat)
         break;
@@ -252,37 +252,37 @@ static void regRstInitPII(struct reg_rst_pars  *pars,
     case 4:                                             // Algorithm 4 : Pure delay fraction 1.401 - 2.01 (not dead-beat)
 
         c2 = exp(-pars->period * TWO_PI * clbw3);
-        q1 = 2.0 - a1 + c1 + c2 + d1;
+        q1 = 2.0 - (double)a1 + c1 + c2 + d1;
 
-        pars->rst.r[0] = (4*a1 + 2*c1 + 2*c2 + 2*d1 + d2 + 3*a1*c1 + 3*a1*c2 + 3*a1*d1 + 2*a1*d2 + c1*c2 + c1*d1 + c2*d1 + 
+        pars->rst.r[0] = (4*(double)a1 + 2*c1 + 2*c2 + 2*d1 + d2 + 3*a1*c1 + 3*a1*c2 + 3*a1*d1 + 2*a1*d2 + c1*c2 + c1*d1 + c2*d1 +
                           2*a1*c1*c2 + 2*a1*c1*d1 + a1*c1*d2 + 2*a1*c2*d1 + a1*c2*d2 - c1*c2*d2 + a1*c1*c2*d1 + 3)/(b0_b1*(a1 + 1)*(a1 + 1)) + 
                           (b1*(c1 + 1)*(c2 + 1)*(d1 + d2 + 1))/(b0_b1*b0_b1*(a1 + 1)) - 
                           (a1*(a1 - c1)*(a1 - c2)*(a1*a1 - d1*a1 + d2))/((a1 + 1)*(a1 + 1)*(b1 - a1*b0));        
 
-        pars->rst.r[1] = (c1*d2 - c2 - d1 - c1 + c2*d2 + 3*a1*a1*c1 + 3*a1*a1*c2 + 3*a1*a1*d1 + 2*a1*a1*d2 + 4*a1*a1 + c1*c2*d1 + 2*c1*c2*d2 + 
+        pars->rst.r[1] = ((double)c1*d2 - c2 - d1 - c1 + c2*d2 + 3*a1*a1*c1 + 3*a1*a1*c2 + 3*a1*a1*d1 + 2*a1*a1*d2 + 4*a1*a1 + c1*c2*d1 + 2*c1*c2*d2 +
                           2*a1*a1*c1*c2 + 2*a1*a1*c1*d1 + a1*a1*c1*d2 + 2*a1*a1*c2*d1 + a1*a1*c2*d2 + a1*a1*c1*c2*d1 - 2)/(b0_b1*(a1 + 1)*(a1 + 1)) + 
                          (2*a1*(a1 - c1)*(a1 - c2)*(a1*a1 - d1*a1 + d2))/((a1 + 1)*(a1 + 1)*(b1 - a1*b0)) + 
                          (b1*(a1 - 1)*(c1 + 1)*(c2 + 1)*(d1 + d2 + 1))/(b0_b1*b0_b1*(a1 + 1));
 
-        pars->rst.r[2] = (a1*(2*a1 + a1*c1 + a1*c2 + a1*d1 - a1*a1 - c1*c2*d2)*b0*b0 + 
+        pars->rst.r[2] = ((double)a1*(2*a1 + a1*c1 + a1*c2 + a1*d1 - a1*a1 - c1*c2*d2)*b0*b0 +
                           a1*(4*a1 - c1 - c2 - d1 + 2*a1*c1 + 2*a1*c2 + 2*a1*d1 + c1*d2 + c2*d2 - 2*a1*a1 + c1*c2*d1 - 2)*b0*b1 - 
                           a1*(2*c1 - 2*a1 + 2*c2 + 2*d1 + d2 - a1*c1 - a1*c2 - a1*d1 + c1*c2 + c1*d1 + c2*d1 + a1*a1 + 3)*b1*b1)/(b0_b1*b0_b1*(b1 - a1*b0));
         
-        q2 = (b1*(c1 + c2 + d1 - c1*d2 - c2*d2 - c1*c2*d1 - 2*c1*c2*d2 + 2) + a1*b1*(2*c1 + 2*c2 + 2*d1 + d2 + c1*c2 + c1*d1 + c2*d1 - c1*c2*d2 + 3))/(b0_b1*(a1 + 1)*(a1 + 1)) + 
+        q2 = ((double)b1*(c1 + c2 + d1 - c1*d2 - c2*d2 - c1*c2*d1 - 2*c1*c2*d2 + 2) + a1*b1*(2*c1 + 2*c2 + 2*d1 + d2 + c1*c2 + c1*d1 + c2*d1 - c1*c2*d2 + 3))/(b0_b1*(a1 + 1)*(a1 + 1)) +
              (b1*b1*(c1 + 1)*(c2 + 1)*(d1 + d2 + 1))/(b0_b1*b0_b1*(a1 + 1)) + 
              (b1*(a1 - c1)*(a1 - c2)*(a1*a1 - d1*a1 + d2))/((a1 + 1)*(a1 + 1)*(b1 - a1*b0));
         
         pars->rst.s[0] = 1.0;
-        pars->rst.s[1] = q1 - 2.0;
-        pars->rst.s[2] = q2 - 2.0*q1 + 1.0;
-        pars->rst.s[3] = q1 - 2.0*q2;
+        pars->rst.s[1] = (double)q1 - 2.0;
+        pars->rst.s[2] = (double)q2 - 2.0*q1 + 1.0;
+        pars->rst.s[3] = (double)q1 - 2.0*q2;
         pars->rst.s[4] = q2;
     
         pars->rst.t[0] = 1.0 / b0_b1;
-        pars->rst.t[1] = (c1 + c2 + d1) / b0_b1;
-        pars->rst.t[2] = (d2 + c1*c2 + c1*d1 + c2*d1) / b0_b1;
-        pars->rst.t[3] = (c1*d2 + c2*d2 + c1*c2*d1) / b0_b1;
-        pars->rst.t[4] = c1*c2*d2 / b0_b1;
+        pars->rst.t[1] = ((double)c1 + c2 + d1) / b0_b1;
+        pars->rst.t[2] = ((double)d2 + c1*c2 + c1*d1 + c2*d1) / b0_b1;
+        pars->rst.t[3] = ((double)c1*d2 + c2*d2 + c1*c2*d1) / b0_b1;
+        pars->rst.t[4] = (double)c1*c2*d2 / b0_b1;
      
         pars->rst.track_delay = 3.0 * pars->period;         // Track delay is about 3 periods (not dead-beat)
         break;
